Make grml-tips start with 1 and add title to Tip #216
I tried to identify some kind of title for each of the Grml Tips with
something like this:

  ❯ grep -e "^-- " -A1 grml_tips | grep -v -- "^--" | cat -n

This attempt did not display the first Grml Tip, so I added the "-- "
before the first line. This had the side effect, that Grml Tips now
start with 1 (instead of 0). I think this is better (unless we're
referencing the Grml Tips from somewhere that I am not aware of).

This change will most probably affect the Grml Tips Hugo template which
parses the grml_tips directly from Github, see:
https://github.com/grml/grml.org/blob/401cb9c1036c1432849ecbc54dcc57dd9027566d/content/tips/index.md?plain=1#L14

This should be changed anyway.

I then noticed that only one tip had no "title": #216. I came up with a
title and, while at it, also fixed some typos in the description.
I also added the tag 'network' for the tip.

The grml_tips list looks like this after the change:

  ❯ grep -e "^-- " -A1 grml_tips | grep -v -- "^--" | cat -n
       1  Configure network:
       2  Deactivate error correction of zsh:
       3  Disable automatic setting of title in GNU screen:
       4  Do not use menu completion in zsh:
       5  Run GNU screen with grml-configuration:
  [...]
     216  Limit SSH connection attempts
  [...]
     271  Install Archlinux using Grml:
     272  Export blockdevices via AoE (ATA over Ethernet):
     273  Access blockdevices via AoE (ATA over Ethernet):
     274  Check notebook's battery status:
     275  Change notebook's screen brightness:

Issues: grml/grml-tips#6, grml/grml-tips#7